+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## 🛑 Breaking changes 🛑
 
 - `memcachedreceiver`: Update metric names (#6594)
+- `memcachedreceiver`: Fix some metric units and value types (#6895)
 - `sapm` receiver: Use Jaeger status values instead of OpenCensus (#6682)
 - `jaeger` receiver/exporter: Parse/set Jaeger status with OTel spec values (#6682)